refactor(runtime): migrate more stdlib overrides to source patches
