+++ runtime/_patch/internal/runtime/maps/typestring_go125.go
@@ -1,0 +1,13 @@
+//go:build go1.25 && !go1.26
+
+package maps
+
+//llgo:skip typeString
+
+import (
+	"internal/abi"
+	_ "unsafe"
+)
+
+//go:linkname typeString github.com/goplus/llgo/runtime/abi.(*Type).String
+func typeString(typ *abi.Type) string
