+++ runtime/_patch/internal/runtime/sys/intrinsics.go
